Cleanup
Looking at the assignment docs, it seems less convoluted to use a
second script to validate the correctness of the algorithm.

If we want to be really clean, we could package the algorithms as
modules, or we could just dump the functions into both files.  Modules
may create slightly more complexity, so given the time constraint, we
may want to just do it the ugly way and clean up later if we have time.